pyproject.toml
[tool.poetry]
name = "unbiased factorization machines"
version = "0.1.0"
description = ""
authors = ["tatsuki takahashi"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.24.3"
pandas = "^2.0.1"
scikit-learn = "^1.2.2"
tqdm = "^4.65.0"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
hydra-core = "^1.3.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
jupyterlab = "^3.1.16"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py39']
line-length = 88
